Finalized Basic Player Movement
Tweaks will be applied later
