+++ CodeFactory/Assembly-CSharp.csproj
@@ -62,8 +62,11 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Assets\Scripts\PlayerMovement.cs" />
+    <Compile Include="Assets\Scripts\PlayerController.cs" />
+    <Compile Include="Assets\Scripts\PlayerMovementFree.cs" />
+    <Compile Include="Assets\Scripts\PlayerMovementLock.cs" />
     <Compile Include="Assets\Scripts\SettingsController.cs" />
+    <Compile Include="Assets\WaitForKey.cs" />
     <Reference Include="Unity.Timeline.Editor">
       <HintPath>C:/Users/macat/Documents/GitHub/CodeFactory/CodeFactory/Library/ScriptAssemblies/Unity.Timeline.Editor.dll</HintPath>
     </Reference>
