+++ CodeFactory/Assets/Scripts/PlayerMovementFree.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0bdba4c393a8ec447bf64cdf24b82b2b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
