+++ CodeFactory/Assets/Scripts/PlayerMovementLock.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a0151adaeba56c343845db19d15698e7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
