+++ CodeFactory/Assets/WaitForKey.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4c4395287f06e704ea0a4b4da65e4a65
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
