+++ CodeFactory/Library/StateCache/SceneView/822906-mainStage.json
@@ -1,1 +1,1 @@
-﻿{"cameraMode":{"drawMode":0,"name":"Shaded","section":"Shading Mode"},"sceneLighting":true,"audioPlay":false,"sceneViewState":{"showFog":true,"showMaterialUpdate":false,"showSkybox":true,"showFlares":true,"showImageEffects":true,"showParticleSystems":true},"in2DMode":false,"pivot":{"x":0.0,"y":4.497381687164307,"z":-4.653692245483398},"rotation":{"x":-0.13313829898834229,"y":0.7042077779769898,"z":-0.13713112473487855,"w":-0.6838079690933228},"size":17.942371368408204,"orthographic":false}
+﻿{"cameraMode":{"drawMode":0,"name":"Shaded","section":"Shading Mode"},"sceneLighting":true,"audioPlay":false,"sceneViewState":{"showFog":true,"showMaterialUpdate":false,"showSkybox":true,"showFlares":true,"showImageEffects":true,"showParticleSystems":true},"in2DMode":false,"pivot":{"x":-0.018221549689769746,"y":4.503822326660156,"z":-4.650689125061035},"rotation":{"x":-0.2321743667125702,"y":-0.0052588265389204029,"z":0.001240634243004024,"w":-0.9726822376251221},"size":0.7877798676490784,"orthographic":false}
